PatternRecognition/requirements.txt
tensorflow>=2.4.0
numpy>=1.19.2
matplotlib>=3.3.2
pillow>=8.0.1
scikit-learn>=0.23.2
opencv-python>=4.4.0
pandas 
seaborn
skimage